Give each major Rails change their own app files
