+++ sentry-rails/spec/support/test_rails_app/apps/6-1.rb
@@ -1,6 +1,37 @@
+require "active_storage/engine"
+require "action_cable/engine"
+
 ActiveRecord::Base.establish_connection(adapter: "sqlite3", database: "db")
 
 ActiveRecord::Schema.define do
+  create_table "active_storage_attachments", force: :cascade do |t|
+    t.string "name", null: false
+    t.string "record_type", null: false
+    t.integer "record_id", null: false
+    t.integer "blob_id", null: false
+    t.datetime "created_at", null: false
+    t.index ["blob_id"], name: "index_active_storage_attachments_on_blob_id"
+    t.index ["record_type", "record_id", "name", "blob_id"], name: "index_active_storage_attachments_uniqueness", unique: true
+  end
+
+  create_table "active_storage_blobs", force: :cascade do |t|
+    t.string "key", null: false
+    t.string "filename", null: false
+    t.string "content_type"
+    t.text "metadata"
+    t.string "service_name"
+    t.bigint "byte_size", null: false
+    t.string "checksum", null: false
+    t.datetime "created_at", null: false
+    t.index ["key"], name: "index_active_storage_blobs_on_key", unique: true
+  end
+
+  create_table "active_storage_variant_records", force: :cascade do |t|
+    t.integer "blob_id", null: false
+    t.string "variation_digest", null: false
+    t.index ["blob_id", "variation_digest"], name: "index_active_storage_variant_records_uniqueness", unique: true
+  end
+
   create_table :posts, force: true do |t|
   end
 
@@ -11,10 +42,15 @@ end
 
 class ApplicationRecord < ActiveRecord::Base
   self.abstract_class = true
+
+  include ActiveStorage::Attached::Model
+  include ActiveStorage::Reflection::ActiveRecordExtensions
+  ActiveRecord::Reflection.singleton_class.prepend(ActiveStorage::Reflection::ReflectionExtension)
 end
 
 class Post < ApplicationRecord
   has_many :comments
+  has_one_attached :cover
 end
 
 class Comment < ApplicationRecord
@@ -29,6 +65,20 @@ class PostsController < ActionController::Base
 
   def show
     p = Post.find(params[:id])
+
+    render plain: p.id
+  end
+
+  def attach
+    p = Post.find(params[:id])
+
+    attach_params = {
+      io: File.open(File.join(Rails.root, 'public', 'sentry-logo.png')),
+      filename: 'sentry-logo.png',
+      service_name: "test"
+    }
+
+    p.cover.attach(attach_params)
 
     render plain: p.id
   end
@@ -75,6 +125,9 @@ class HelloController < ActionController::Base
 end
 
 def make_basic_app
+  ActionCable::Channel::Base.reset_callbacks(:subscribe)
+  ActionCable::Channel::Base.reset_callbacks(:unsubscribe)
+
   app = Class.new(TestApp) do
     def self.name
       "RailsTestApp"
@@ -86,6 +139,8 @@ def make_basic_app
   app.config.logger = Logger.new(nil)
   app.config.eager_load = true
 
+  app.config.active_storage.service = :test
+
   app.routes.append do
     get "/exception", :to => "hello#exception"
     get "/view_exception", :to => "hello#view_exception"
@@ -93,7 +148,11 @@ def make_basic_app
     get "/not_found", :to => "hello#not_found"
     get "/world", to: "hello#world"
     get "/with_custom_instrumentation", to: "hello#with_custom_instrumentation"
-    resources :posts, only: [:index, :show]
+    resources :posts, only: [:index, :show] do
+      member do
+        get :attach
+      end
+    end
     get "500", to: "hello#reporting"
     root :to => "hello#world"
   end
